Improve Rust crate directory tree
Split examples module on integration test and example. Moving code into
a proper place of the crate's directory tree. Thus standard locations
for examples and integration tests are added.
