--- lib/src/metta/examples/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod types;
